requirements.txt
requests
beautifulsoup4
pandas
numpy
matplotlib
seaborn
mysql-connector-python
python-dotenv
prophet